+++ laktory/_testing/stockprices.py
@@ -2,20 +2,6 @@ import os
 
 import pandas as pd
 import polars as pl
-from pyspark.sql import SparkSession
-
-spark = (
-    SparkSession.builder.appName("UnitTesting")
-    .config("spark.jars.packages", "io.delta:delta-spark_2.12:3.2.0")
-    .config("spark.sql.extensions", "io.delta.sql.DeltaSparkSessionExtension")
-    .config(
-        "spark.sql.catalog.spark_catalog",
-        "org.apache.spark.sql.delta.catalog.DeltaCatalog",
-    )
-    .getOrCreate()
-)
-spark.conf.set("spark.sql.session.timeZone", "UTC")
-
 
 # --------------------------------------------------------------------------- #
 # DataFrames                                                                  #